+++ lib/generators/rlm/install/templates/rlm_trace.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class RlmTrace < ApplicationRecord
+  self.table_name = "rlm_traces"
+end